requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sentence-transformers==5.1.2
huggingface-hub==0.36.0
razorpay==1.4.2
python-dotenv==1.0.0
bcrypt==4.0.1
pydantic==2.5.0
torch==2.9.0
numpy==1.26.4
email-validator==2.0.0
python-multipart==0.0.6
gTTS==2.5.4